chore(ci): add benchmarks workflow
Benchmarks would be run in GitHub Actions each time a commit is
pushed into main and can also be triggered on demand on any branch.
Benchmarks are executed in two flavors, one with AVX512 CPU
instructions support and one without it.
Results are parsed and then sent to Slab, Zama's CI bot, to be
stored in database.